requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-mysqldb==2.0.0
PyMySQL==1.1.1
werkzeug==3.0.3
requests==2.31.0
PyJWT==2.8.0
Pillow==10.1.0
boto3==1.34.0
botocore==1.34.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
python-dotenv==1.0.0
flask-cors==4.0.0
gunicorn==21.2.0

# Dev tooling
basedpyright==1.31.3

